refactor: refactor plane mesh
