+++ packages/core/src/geometry-shape/index.ts
@@ -1,5 +1,4 @@
 export { CylinderGeometry } from "./CylinderGeometry";
-export { PlaneGeometry } from "./PlaneGeometry";
 export { ScreenQuadGeometry } from "./ScreenQuadGeometry";
 export { ShapeGeometry } from "./ShapeGeometry";
 export { PrimitiveMesh } from "./PrimitiveMesh";
